refactor(Spring) : 타입 수정
- spotless 적용

+++ deepline-backend/src/main/java/crushedtaro/deeplinebackend/domain/organization/dto/DepartmentDTO.java
@@ -1,3 +1,3 @@
 package crushedtaro.deeplinebackend.domain.organization.dto;
 
-public record DepartmentDTO(Long deptCode, String deptName) {}
+public record DepartmentDTO(int deptCode, String deptName) {}
